python/niualcoholbottle01/requirements.txt
line-bot-sdk
flask
PyMySQL==0.9.2
demjson==2.2.4